refactor admin pages into component folders
