+++ client/src/Admin/pages/Clients/components/ClientList.tsx
@@ -1,0 +1,79 @@
+import { useEffect, useRef, useState } from 'react'
+import { Link } from 'react-router-dom'
+import { Client } from './types'
+
+export default function ClientList() {
+  const [items, setItems] = useState<Client[]>([])
+  const [search, setSearch] = useState('')
+  const [page, setPage] = useState(0)
+  const [hasMore, setHasMore] = useState(true)
+  const loader = useRef<HTMLDivElement | null>(null)
+
+  useEffect(() => {
+    setItems([])
+    setPage(0)
+    setHasMore(true)
+  }, [search])
+
+  useEffect(() => {
+    load()
+  }, [page, search])
+
+  function load() {
+    fetch(`http://localhost:3000/clients?search=${encodeURIComponent(search)}&skip=${page * 20}&take=20`)
+      .then((r) => r.json())
+      .then((data: Client[]) => {
+        setItems((prev) => {
+          const next = page === 0 ? data : [...prev, ...data]
+          const seen = new Set<number>()
+          return next.filter((c) => {
+            if (c.id === undefined) return false
+            const exists = seen.has(c.id)
+            seen.add(c.id)
+            return !exists
+          })
+        })
+        if (data.length < 20) setHasMore(false)
+      })
+  }
+
+  useEffect(() => {
+    const obs = new IntersectionObserver((entries) => {
+      if (entries[0].isIntersecting && hasMore) {
+        setPage((p) => p + 1)
+      }
+    })
+    if (loader.current) obs.observe(loader.current)
+    return () => {
+      if (loader.current) obs.unobserve(loader.current)
+    }
+  }, [hasMore])
+
+  return (
+    <div className="p-4">
+      <h2 className="text-xl font-semibold mb-2">Clients</h2>
+      <div className="flex items-center gap-2 mb-4">
+        <input
+          value={search}
+          onChange={(e) => setSearch(e.target.value)}
+          placeholder="Search by name or number"
+          className="flex-1 border p-2 rounded"
+        />
+        <Link to="new" className="bg-blue-500 text-white px-3 py-1 rounded text-sm">
+          New
+        </Link>
+      </div>
+      <ul className="divide-y">
+        {items.map((c) => (
+          <li key={c.id} className="py-2">
+            <Link to={String(c.id)} className="block">
+              <div className="font-medium">{c.name}</div>
+              <div className="text-sm text-gray-600">{c.number}</div>
+            </Link>
+          </li>
+        ))}
+      </ul>
+      <div ref={loader} className="h-5" />
+    </div>
+  )
+}
