+++ client/src/Admin/pages/Clients/components/types.ts
@@ -1,0 +1,6 @@
+export interface Client {
+  id?: number
+  name: string
+  number: string
+  notes?: string
+}
