+++ client/src/Admin/pages/Employees/components/types.ts
@@ -1,0 +1,6 @@
+export interface Employee {
+  id?: number
+  name: string
+  number: string
+  notes?: string
+}
